Add context and propagate settings to plugins
